need to clarify this policy is only applicable to Mule 4.x and above
Added following as this policy is only applicable to Mule 4.x but customers may think they are Using API Manager v2 thus they can use it as well. 
This policy is applicable to Mule 4.x and above only and you can see it for API Manager instances that have the checkbox 'Check this box if you are managing this API in Mule 4 or above' checked under API instance setting page.